Moving decorator to its own file
