+++ src/WidgetBase.ts
@@ -17,8 +17,7 @@ import {
 	PropertyChangeRecord,
 	WidgetBaseInterface,
 	WidgetConstructor,
-	WidgetProperties,
-	WNode
+	WidgetProperties
 } from './interfaces';
 import WidgetRegistry, { WIDGET_BASE_TYPE } from './WidgetRegistry';
 
@@ -102,35 +101,6 @@ export function handleDecorator(handler: (target: any, propertyKey?: string) => 
 			handler(target, propertyKey);
 		}
 	};
-}
-
-function propertyPropagator(this: WidgetBase<any>, node: DNode) {
-	const propagatedProperties: string[] = (<any> this).getDecorator('propagateProperty').filter((propertyName: string) => {
-		return propertyName in this.properties;
-	});
-
-	decorate(node, (wNode: WNode) => {
-		propagatedProperties
-			.filter(propertyName => !(propertyName in wNode.properties))
-			.forEach(propertyName => {
-				(<any> wNode.properties)[propertyName] = this.properties[propertyName];
-			});
-	}, isWNode);
-
-	return node;
-}
-
-const propagateInstanceMap = new WeakMap<WidgetBase<any>, boolean>();
-
-export function propagateProperty(propertyName: string) {
-	return handleDecorator(target => {
-		target.addDecorator('propagateProperty', propertyName);
-
-		if (!propagateInstanceMap.has(target)) {
-			afterRender(propertyPropagator)(target);
-			propagateInstanceMap.set(target, true);
-		}
-	});
 }
 
 /**
