+++ src/mixins/Registry.ts
@@ -1,12 +1,8 @@
 import { includes } from '@dojo/shim/array';
+import { propagateProperty } from '../decorators/propagateProperty';
+import { PropertyChangeRecord, PropertiesChangeEvent, Constructor, WidgetProperties } from '../interfaces';
 import WidgetRegistry from '../WidgetRegistry';
-import { WidgetBase, onPropertiesChanged, diffProperty, propagateProperty } from './../WidgetBase';
-import {
-	PropertyChangeRecord,
-	PropertiesChangeEvent,
-	Constructor,
-	WidgetProperties
-} from '../interfaces';
+import { WidgetBase, onPropertiesChanged, diffProperty } from '../WidgetBase';
 
 export interface RegistryMixinProperties extends WidgetProperties {
 	registry: WidgetRegistry;
